Changed chat and message IDs to long's to enable support for supergroups

+++ Roboto/Modules/ModuleCode/mod_steam_steamapi.cs
@@ -24,7 +24,7 @@ namespace Roboto.Modules
         }
 
 
-        public static mod_steam_player getPlayerInfo(long playerID, int chatID)
+        public static mod_steam_player getPlayerInfo(long playerID, long chatID)
         {
             NameValueCollection pairs = new NameValueCollection();
             pairs["steamids"] = playerID.ToString();
